+++ Dashboard/Subscriber/update.jsx
@@ -1,0 +1,9 @@
+const Update = () => {
+  return (
+    <div>
+      <h1>This is the update section</h1>
+    </div>
+  );
+};
+
+export default Update;
